requirements.txt
liquid-dsp
matplotlib
numpy
pybind11[global]
